[chore] Move resource test to service/internal/resource
